:sparkles: feat: 인기 검색어, 카테고리 구현

+++ src/main/kotlin/com/team4/moviereview/domain/search/dto/SearchWordResponse.kt
@@ -2,5 +2,5 @@ package com.team4.moviereview.domain.search.dto
 
 data class SearchWordResponse(
     val keyword: String,
-    val count: Int
+    val count: Long
 )
